+++ Blog2.0/blog2.0/src/app/components/about-page/picture.model.ts
