+++ Blog2.0/blog2.0/src/app/components/posts/picture.model.ts
