+++ Blog2.0/blog2.0/src/app/components/posts/post.model.ts
@@ -1,4 +1,4 @@
-import { Picture }  from './picture';
+import { Picture }  from './picture.model';
 
 export class Post {
     id: number;
