+++ Blog2.0/blog2.0/src/app/components/posts/posts.component.ts
@@ -1,12 +1,12 @@
 import { Component, OnInit, HostListener } from '@angular/core';
 import * as _ from 'lodash';
 
-import { PostService } from '../../services/post.service';
-import { PostModelService } from '../../services/postModelService'
-import { Picture as PictureModelService} from '../../services/picture';
+import { PostService } from '../../services/post/post.service';
+import { PostModelService } from '../../services/post/postModelService.model'
+import { Picture as PictureModelService} from '../../services/post/picture.model';
 
-import { Post } from './post';
-import { Picture } from './picture';
+import { Post } from './post.model';
+import { Picture } from './picture.model';
 
 import { ContentLayout } from '../../modules/contentLayout.module';
