+++ Blog2.0/blog2.0/src/app/services/post/postModelService.model.ts
@@ -1,4 +1,4 @@
-import { Picture }  from './picture';
+import { Picture }  from './picture.model';
 
 export class PostModelService {
     Id: number;
